requirements.txt
locust>=1.0.0,<2
PyYAML>=5.3.1
